Update README.md: Getting started first section
To start with the project an overview of the project should be given.
Therefore the section: Getting Started was added. It includes a short
description about the goal of the project, the context in which this
project is started as well as its first goal.
It also includes why we are changing the project structure from the old
OS project to the new OS project.