Change foldername from OmnAIView->angular-frontend
Problem:

As the project grows the project will not only have a angular frontend
but also a electron app as well as a possible android app. Therefore
the projects should be seperated by their frontend type as they are a
part of OmnAIView.

Changes:

The name for the angular frontend setup is now called: angular frontend
instead of OmnAIView