Using angular in electron app
As we are not using webpack and electron independend from the angular
application, webpack as well as the html files from the electron project
are no longer used. Instead a main.ts is used that will only render
the angular project. Therefore the angular project is included into the
electron project.

Important!: The tsc version was changed to a higher version because the
older version is no longer supported.