Feature: Starting Backend on a free port
Problem: For v0.5.0 to start the backend with a websocket connection
a port has to be given in the CLI.

Solution: Searching for a free port by testing a free port on the
clients device instead of setting a hardcoded port. Giving the port as
a parameter into spawn when starting the backend.

Changes:
- Implementation of the above behavior
- Changing the app version to 0.2.0 in the package.json as this is a new
  feature