Fix: Receiving data from backend
Problem : Data from the OmnAIScope was not drawn in the graph

Solution: The Backend expects the UUID from the devices that should send
data as a string . The UUID was send as a JSON string which means it had
quotation marks around it. The Backend expects a normal string without
quotation marks. This was fixed by this commit.