Feature : Toolbar for starting a measurement
What ?
A toolbar where the user can start, stop and delete a measurement.
The toolbar uses the settings the user chose or the default settings.

Why ?
To make the measurement easy accessible by one click with the default
settings. The user wants to stop the measurement and also be able to
delete the measurement .

How ?
The Measurement component receives the informations from the settings
service and starts/Stops/deletes the measurement via the
data-source-selection-service.