Fix: Button reset when deleting measurement
Problem: When pressing delete the buttons do not reset

Issue: The button array was a reference of the buttonsDefault. Therefore
changing the buttons also changed the default. Reseting would reset the
buttons to their current state.

Changes: Instead of using a reference the buttons are a copy of the
default buttons. Reseting the buttons now actually resets them to their
default value.