Showcase minimal datasource integration
This commit intends to show how to use the current pattern and what
files to adapt to make a source useable.

The example is the integration of the DevDataServer as a new Datasource.
This is the minimal needed integration of the DevDataServer as a new
DataSource.

Steps taken:
1. Set DevdataServer as a Datasource in the availableSources
in the DataSourceSelectionService

2. Setup of a minimal service that implements the DataSource interface
   This contains:
   - a connect function
   - a data signal

Note: The service only provides the interface implementation not the functionality.