Showcase minimal datasource functionality
This commit intends to show how to setup the functionality for a new
datasource, so data is visualized in the graph.

The example is the implementation of basic functionality for the DevDataServer.

1. getDevices : gets the list of possible devices from the server
   provides the list in a devices signal
   This is needed otherwise the message the websocket sends would be
   empty
2. connect : opens the websocket connection to the server
   receives and parses the data. Appends the data to the existingData
   and updates the data signal .

Note: The functionality cant be used as no UI is provided to get the
devices from the server