Showcase minimal datasource UI setup
This commit intends to show how to setup the minimal UI needed to
interact with a new integrated datasource.

The example is the UI for the DevDataServer

Problem : Without Information about the devices the Dataserver provides
it is not possible to receive data from the server.

Solution: Implement temporary button: Load Test Server Devices list

Intended Userflow:
1. Click on Load Test Server Devices list button
2. Devices should be shown
3. Click on Start data
4. Choose the test data server
5. Click on Start
6. The data should be shown in the graph