Minimise data-copying, by using in-place map updates
The map itself is wrapped in an extra object, because signals only update, when the object in the signal is a different object. In-Place updates of a Data-Structure inside a signal don't trigger the signal change code.

Initially I profiled this, by checking how long the websocket handler took to process. (performance.now() at the start and end, then console.log() the difference)
I found this way of doing things to be way faster, in regard to websocket handling.