Only update every-so-often on websocket messages
When the websocket is hit with a ton of small messages, this helps to not over-strain the graph-renderer too much, by re-rendering the graph on every websocket message.

With the last commit, I found that large messages over the websocket are handled very well now.
However, when receiving a ton of small messages, this tends to not be the case.
In order to stop an alternate implementation to make the complete webpage unusable some rate-limits were introduced to not update the graph every-time (especially with how costly graph updates are).