Showcase minimal datasource configuration
This commit intends to show how to implement minimal datasource
configuration for a new integrated datasource.
The example is the UUID configuration of the DevDataServer.

Changes:
- Implement popupModal that starts from within the connect function of the
dataserver.
- Configure message that is send via websocket connection

Intended Userflow:

- Load Test data via the load Test server devices list button
- Click on start data
- Select Testserver as source
- Click Start
- Configure your Websocket by selecting the prefered UUID
- Click on Connect
You should be able to see the waveform from the sinus or rectangular
generator

Note: This example does not provide a stop or delete possibility
therefore connection is only possible once .