Doc: Init VISION.md
Idea: Currently many questions arise what the goal of the frontend is,
which features have a high priority, which features are even needed.

To present the overall goal and the specific goals for the next months
the VISION.md is created.
For the first time it gives an idea what the AI-Gruppe goals for this
application are.
I would suggest that important design choices from the
discussion section and the monthly meeting are documented here
as this project and the vision evolves further.