feat(xDomain):improve x-axis format with absolute and relative timestamp options (#58)
* Original author is @giayyuk

* chore: cleanup imports

remove unused imports and organize them alphabetically.

* feat(x-domain modi): allow absolute and relative timestamps

add state and UI for switiching between absolute and relative timestamps
on the x-axis.

Notes: The acutal formatting is not done in this commit.

* fix(xDomain): absolute values only

As creating a Dateobject with 0 as a value will create
Date information that are incorrect and might lead to wrong comparisons,
undo the relative changes done to the domain updates

* feat(xDomain):add default x-Domain values

To have a coherent setting of the x-Domain in case no values are
provided, add a constant that contains useful values and allows
for easier changes of the default domain

* clean(xDomain): remove unused code

* feat(xDomain):add formatter for different scenarios

Depending on the selected mode (absolute or relative),
set the formatter for the xAxis.

Notes: tested with random data and works. No units added currently

* clean(xDomain):add type for xAxisMode

Will improve type information in case more x Domain options are implemented.

* docs(xDomain):describe x-axis updates

* chore(test):fix testbed setups

add missing providers to testbeds, no functional
changes but now the creation testcases are working

---------

Signed-off-by: Tabea <tabea.roethemeyer@gruppe.ai>
Co-authored-by: Lukas Adrian Kron <l.kron@lugges.eu>