Add privacy-default (#82)
* Add privacy-default

* Update CHANGELOG.md

Angular (a product of Google) asks to collect usage statistics. You have to approve or deny it in the CLI.
I think this could be an easy feature to improve privacy and by default colleting as few data as possible (even if it is by a third party), and thus denying by default.