catch electron-squirrel setup error (#80)
This hotfix addresses a critical issue that caused the Electron build to fail due to a missing dependency (electron-squirrel-startup). Additionally, adjustments were made to the CI workflow to reflect recent script changes.
Problem: The Electron process crashed during startup when electron-squirrel-startup was not available, due to an unguarded require(...) statement.

Signed-off-by: Lukas Adrian Kron <l.kron@lugges.eu>