doc(enhancement): Improve PR template. (#93)
* doc(enhancement): PR template clearer

Improve PR template.

Problem: The current template has to much default text which makes it
hard to see the important parts of the description directly.
Using details drop down and less but more precise descriptions should improve this.

* doc(enhancement): Check Changelog entry in PR

Add the documententation of a change in the CHANGELOG.md as a
checkpoint for the PR template.

* Update Changelog.md

* doc(enhancement): PR Template update

Update of PR template with requested changes.

Comments instead of drop down boxes.
Checkbox for title

Co-authored-by: franzwil

* Update Changelog

* Update .github/PULL_REQUEST_TEMPLATE.md

Co-authored-by: Tabea <tabea.roethemeyer@gruppe.ai>
Signed-off-by: AKMaily <83161524+AKMaily@users.noreply.github.com>

* fix(doc): Set checklist at top of PR Template

Set checklist and type of change at the top of PR template.

This structure improves readbility for the contributer as they see what
they should have done before starting a PR.

---------

Signed-off-by: AKMaily <83161524+AKMaily@users.noreply.github.com>
Co-authored-by: Tabea <tabea.roethemeyer@gruppe.ai>