feat(setup): Add CI workflow for release (#123)
* feat(setup): CI workflow for release

Add CI workflow for automatic release on new merges on master.

Versioning or skipping this workflow is handled via PR labels.

* fix(setup): Use prerelease tags correctly

The prerelease tags were not added to the tag name correctly.

Adding the pre- to the version as well when a prerelease tag is used
should fix this.

* doc(CI-Release): Add CI-Release documentation

Add documentation of labels needed and steps for the CI workflow in the
release.yml .
Add Update to Changelog.mc