Fix: Versionnumber + Changelog  (#139)
* fix(Version): Add manual version

Add manual version in package.json files as the CI for releases is
currently not working and the version was not updated for the last
versions that were released. As long as no fix for the CI build is found
the version should be updated manually in the PR.

* fix(doc): Change Changelog to fit versions

Change changelog to fit the version numbers and the keepachangelog
format.

--> Add changes made in a specific version to the versions section