fix(zoom): Add correct startpoint for relative time  (#141)
* fix(zoom): Add correct startpoint for rel time

Fix zoom behavior for relative time by using correct start time for
x-axis-formatter-util function .

Problem: Previously the relative time was calculated from the current
date as well as the start date of the domain.
Zooming changes the start data of the domain and therefore the relative
time distance between the current date and the start date.
This lead to wrongly calculated relative times when zooming.

Solution: Receive minTimestamp reference when a measurement starts. Use
reference to calculate relative time to.

* Update package.json to v1.7.1