feat(zoom): Add select to zoom axis separatly
Add checkboxes X-Zoom and Y-Zoom to enable axis zoom for a selected
axis.
Add selection as input for the graph-zoom directive. OnZoom() uses the
input to set transform and axis with setZoom(). Update setZoom to separate
betwenn xZoom and yZoom. XScale() YScale() signals now react to zoomX
and zoomY instead of a shared zoom.
Add comments for documentation.