feat(survey): Add survey tool to graph (#145)
* feat(survey): Add survey tool to graph

Add a survey tool to the graph component. Enabling the tool allows to
measure the distance between two points in the graph.

Add a graph-survey component for the button and functions. The component
relies on the graph-cursor directive.
Add the graph survey component and necessary html to the
graph.component.html to show crosshairs in the graph where the user
clicks to survey.

* Update package.json to v1.10.0