refactor(graph): Refactore graph and main comp (#147)
Refactor graph component into a toolbar, graph and main component.
Previously the graph held all components which were shown in the main
page. This lead to a rapidly expanding graph component.
Separating the component in the graph html into a toolbar component, the
main graph component and a main component should minimize this issue.
The main component holds the main sections of the main page which are
the toolbar, the graph and currently other components that have features
but dont fit in a specific category yet.