feat(setup): Add setup loading multiple servers
Previously only one server for each type was set. Therefore only the
local server could be used in the application.
Add setup to load multiple servers, poll devices every 15 seconds from
the servers and update control panel with information about servers and
devices.

Delete all code that is no longer needed for device polling.

Note: For this to work the port of the omnaiscope backend was set to
8080 again, this will be improved in one of the following commits.

Architectur:
A server is defined as :

Server { ip: string, port:number, name: string, type: ServerType}

Available ServerTypes are :
random, omnaiscope

It is expected that all servers loaded are fullfilling on of the
ServerType interfaces.

Servers are stored in a local file server-config.json. File is loaded
via the server.service function and stored in a serverlist signal.
The devicelist.service calls the server.service function on
construction. Then starts to call getDevices for the specific server
type every 15 seconds.
Server information and device information is stored and updated via the
servercards computed signal for the UI.
Servers and devices are shown in the control panel, using the
servercards via property binding, updating with
serverlist or device changes.